pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zarr-cloud-native-format"
dynamic = ["version"]
description = 'Zarr Cloud-Native Workflow for Water Body Detection'

readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Fabrice Brito", email = "info@terradue.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "odc-stac==0.4.0",
  "ipykernel",
  "cwltool",
  "matplotlib==3.10.6",
  "loguru==0.7.3",
  "zarr",
  "matplotlib",
  "graphviz",
  "pillow",
  "cwl-loader>=0.14.0",
  "plantuml",
  "cwl2puml>=0.0.30.0"
]

[project.urls]
Documentation = "https://eoap.github.io/zarr-cloud-native-format/"
Issues = "https://github.com/eoap/zarr-cloud-native-format/issues"
Source = "https://github.com/eoap/zarr-cloud-native-format"

[tool.hatch.version]
path = "zarr_cloud_native_format/__about__.py"

[tool.hatch.build.targets.wheel]
packages = ["zarr_cloud_native_format"]

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
